Update the basic example name, add exception handling, and tensor slicing. (#1072)
